[cleanup] Remove no longer maintained "XText" projects
Flow Designer has been used as one of the examples for writing "Xtext /
Sirius - Integration White Paper" [1] in November 2017. These projects
are no longer maintained since this date. There is no short or medium
term objective on this subject. To clarify the readability and
maintenance of the Flow Modeler, it was decided to delete all the
corresponding projects.

[1] https://www.obeodesigner.com/en/white-paper-xtext-sirius