[cleanup] Remove wizard and editor duplicated entries
This commit removes one of the two wizards to create a Flow Model. It
removes the entry that is under "Example EMF Model Creation Wizards".
Only the entry under "Examples/Obeo Designer" is kept.

This commit also removes the duplicated editor. This avoids to have two
entries named "Flow Model Editor" when using "Open With" contextual menu
on a "flow" file.